requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
binance-futures-connector==4.1.0
python-dotenv==1.0.1
